+++ EasySaver.BitmapFile/EasySaver.BitmapFile.cs
@@ -47,7 +47,7 @@ namespace EasySaver.BitmapFile
         private readonly static string s_defaultFileName = "";
 
         // Default value if NamingFormat is not provided.
-        private readonly static NamingFormat s_defaultNamingFormat = NamingFormat.DateTime;
+        private readonly static NamingFormat s_defaultNamingFormat = NamingFormat.LongDateTime;
 
         // Default value when using Save(Bitmap bitmap, string fileName) and SaveToFolder(Bitmap bitmap, string folderName, string fileName). Since there is only NamingFormat option available to use by design, this value is internally provided for shorter method usage.
         private readonly static NamingFormat s_onlyBitmapProvidedNamingFormat = NamingFormat.Custom;